// client/src/redux/actions/actionTypes.ts
export enum actionTypes {
    COURSE_INFO_LOADING = 'COURSE_INFO_LOADING',
    COURSE_INFO_SUCC = 'COURSE_INFO_SUCC',
    COURSE_INFO_ERR = 'COURSE_INFO_ERR',
    AUTH_LOADING = 'AUTH_LOADING',
    AUTH_SUCCESS = 'AUTH_SUCCESS',
    AUTH_ERROR = 'AUTH_ERROR',
    SEARCH_SUCC = 'SEARCH_SUCC',
    SEARCH_LOADING = 'SEARCH_LOADING',
    SEARCH_ERR = 'SEARCH_ERR',
    SEARCH_BREAK = 'SEARCH_BREAK',
    POST_REVIEW_LOADING = 'POST_REVIEW_LOADING',
    POST_REVIEW_SUCC = 'POST_REVIEW_SUCC',
    POST_REVIEW_ERR = 'POST_REVIEW_ERR',
    GET_REVIEW_ERR = 'GET_REVIEW_ERR',
    GET_REVIEW_SUCC = 'GET_REVIEW_SUCC',
    GET_REVIEW_LOADING = 'GET_REVIEW_LOADING',
    SEARCH_PROF_REVIEW_SUCC = 'SEARCH_PROF_REVIEW_SUCC',
    SEARCH_PROF_REVIEW_LOADING = 'SEARCH_PROF_REVIEW_LOADING',
    SEARCH_PROF_REVIEW_ERR = 'SEARCH_PROF_REVIEW_ERR',
    SEARCH_PROF_REVIEW_BREAK = 'SEARCH_PROF_REVIEW_BREAK',
    PROF_INFO_LOADING = 'PROF_INFO_LOADING',
    PROF_INFO_SUCC = 'PROF_INFO_SUCC',
    PROF_INFO_ERR = 'PROF_INFO_ERR',
    GET_PROF_REVIEW_ERR = 'GET_PROF_REVIEW_ERR',
    GET_PROF_REVIEW_SUCC = 'GET_PROF_REVIEW_SUCC',
    GET_PROF_REVIEW_LOADING = 'GET_PROF_REVIEW_LOADING',
    POST_PROF_REVIEW_LOADING = 'POST_PROF_REVIEW_LOADING',
    POST_PROF_REVIEW_SUCC = 'POST_PROF_REVIEW_SUCC',
    POST_PROF_REVIEW_ERR = 'POST_PROF_REVIEW_ERR',
    PROF_TAB_INFO_LOADING = 'PROF_TAB_INFO_LOADING',
    PROF_TAB_INFO_SUCC = 'PROF_TAB_INFO_SUCC',
    PROF_TAB_INFO_ERR = 'PROF_TAB_INFO_ERR',
    COURSE_TAB_INFO_LOADING = 'COURSE_TAB_INFO_LOADING',
    COURSE_TAB_INFO_SUCC = 'COURSE_TAB_INFO_SUCC',
    COURSE_TAB_INFO_ERR = 'COURSE_TAB_INFO_ERR',
    GET_REVIEW_APPEND = 'GET_REVIEW_APPEND',
    RESET_REVIEW_MODAL = 'RESET_REVIEW_MODAL',
    RESET_PROF_REVIEW_MODAL = 'RESET_PROF_REVIEW_MODAL',
    GET_PROF_REVIEW_APPEND = 'GET_PROF_REVIEW_APPEND',
    SEARCH_COURSE_REVIEW_LOADING = 'SEARCH_COURSE_REVIEW_LOADING',
    SEARCH_COURSE_REVIEW_SUCC = 'SEARCH_COURSE_REVIEW_SUCC',
    SEARCH_COURSE_REVIEW_ERR = 'SEARCH_COURSE_REVIEW_ERR',
    SEARCH_COURSE_REVIEW_BREAK = 'SEARCH_COURSE_REVIEW_BREAK',
    RESET_AUTH = 'RESET_AUTH',
    DARK_LOADING = 'DARK_LOADING',
    DARK_SUCCESS = 'DARK_SUCCESS',
    DARK_ERROR = 'DARK_ERROR',
    DARK_MODE = 'DARK_MODE',
    AUTH_DISPLAY_CLEAR = 'AUTH_DISPLAY_CLEAR',
    LOG_OUT = 'LOG_OUT',
    VERIFIED = 'VERIFIED',
    VERIFYING = 'VERIFYING',
    RESET_PASS_SUCCESS = 'RESET_PASS_SUCCESS',
    RESET_PASS_ERROR = 'RESET_PASS_ERROR',
    RESET_PASS_LOADING = 'RESET_PASS_LOADING',
    RESET_PASS_DEFAULT = 'RESET_PASS_DEFAULT',

}